ImageTranslator: remove the interface
The ImageTranslator interface brings no benifit.
It's very unlikely that there will by more than one
class implementing it.
